+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - Added `divide(a, b)` function in `functions.py` to perform division.
+- Added `multiply(a, b)` function in `functions.py` to perform multiplication.
 
 
 ## [1.0.0] - 2025-05-11